+++ lib/domain/models/day_summary.dart
@@ -7,7 +7,11 @@ class DaySummary {
 
   final DateTime date;
   final int totalCount;
-  final int doneCount;
+
+  /// Jumlah kredit progress (lihat `Habit.progressCredit`) di hari ini —
+  /// habit yang belum 100% tercapai tetap menyumbang porsinya, bukan cuma
+  /// dihitung kalau sudah penuh.
+  final double doneCount;
 
   double get ratio => totalCount == 0 ? 0 : doneCount / totalCount;
